media-server-buildtools: Don't fail if missing mounts

+++ media-server-buildtools/entrypoint.sh
@@ -2,7 +2,7 @@
 
 # ------------ Shell setup ------------
 
-# Shell options for strict error checking
+# Bash options for strict error checking
 set -o errexit -o errtrace -o pipefail -o nounset
 
 # Trap functions
@@ -18,6 +18,7 @@ trap on_error ERR
 
 
 # ==== Clone or update 'adm-scripts' ====
+
 if [[ -d /adm-scripts/.git ]]; then
     echo "Kurento 'adm-scripts' found: update"
     cd /adm-scripts
@@ -29,18 +30,27 @@ fi
 
 
 # ==== Build the project ====
-rm -rf /workdir
-cp -a /hostdir /workdir
-cd /workdir
+
+if [[ -d /hostdir ]]; then
+    rm -rf /workdir
+    cp -a /hostdir /workdir
+    cd /workdir
+fi
+
 /adm-scripts/kurento-buildpackage.sh "$@"
 # Note: "$@" expands to all quoted arguments, as passed to this script
 
 
 # ==== Finish ====
+
 BASENAME="$(basename "$0")"  # Complete file name
 echo "[$BASENAME] Generated files:"
 # `dh_builddeb` puts the generated .'deb' files in '../'
 find .. -maxdepth 1 -type f ! -name "$BASENAME"
 
 # Get results out from the Docker container
-mv ../*.*deb /hostdir 2>/dev/null || true
+if [[ -d /hostdir ]]; then
+    mv ../*.*deb /hostdir/ 2>/dev/null || true
+else
+    echo "WARNING: No host dir where to put build artifacts"
+fi
